Fix `Sub` to work with mixed-case subtrees
Viper assumes that the top level of the config is lower case. However it
doesn't have this same restriction for sub trees, so you can get in a
sutation where a Viper instance returned by `Sub` has a config with
upper case keys, but all the accessors lowercase the keys.

Also adds a test for the usecase.