Prevent AutomaticEnv from clobbering an explicit mapping
When an explicit environment variable mapping is set for a key, do not
use the automatic mapping. This allows using AutomaticEnv while
specifying a special mapping for a single key.

It's important to both check for the explicit mapping first, and to not
fall back to the automatic mapping when the explicit environment variable
is not set. Otherwise it's very difficult to use AutomaticEnv when you
have a flag that uses a commonly set environment variable, such as
--path. When I map --path to SPECIAL_PATH, it shouldn't fall back to
PATH when SPECIAL_PATH isn't set.