travis: ensure dependencies are encoded in go.mod
Currently a user can add new deps in *.go files that do not exist in the
go.mod/go.sum. This commit adds a small check in CI to ensure that no new
deps were added that aren't in the go.mod/go.sum.

This commit includes a run of `go mod tidy`.